requirements.txt
pytest==7.1.1
allure-pytest==2.9.45